+++ Sources/SwiftGenKit/Parsers/InterfaceBuilder/Platform.swift
@@ -8,10 +8,22 @@ import Foundation
 
 extension InterfaceBuilder {
   enum Platform: String {
-    case tvOS = "AppleTV"
     case iOS = "iOS.CocoaTouch"
     case macOS = "MacOSX.Cocoa"
+    case tvOS = "AppleTV"
     case watchOS = "watchKit"
+
+    static let allCases = [Platform.iOS, .macOS, .tvOS, .watchOS]
+
+    init?(runtime: String) {
+      // files without "trait variations" will have a runtime value of (for example):
+      // "iOS.CocoaTouch.iPad"
+      if let platform = Platform.allCases.first(where: { runtime.hasPrefix($0.rawValue) }) {
+        self = platform
+      } else {
+        return nil
+      }
+    }
 
     var name: String {
       switch self {